+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Features:
 * Add new option `--temperature-unit`
 * Better CPU and Host detection for Android (Android)
 * Support yakuake terminal version & font detection (Terminal, Linux)
+* Add new option `--bright-color` which can be used to disable the default bright color of keys, title and ASCII logo.
 
 Bugfixes:
 * Fix possible hanging (TerminalFont, #493)